refactor(store): split monolith into slices (auth, trips, trip-detail, ui)
Made-with: Cursor
